Rename package `lookup.validate` to `lookup.valrule`
